finishi visualvm
